requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
auth==0.5.3
azure-core==1.30.2
Babel==2.14.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
botbuilder-core==4.15.0
botbuilder-integration-aiohttp==4.15.0
botbuilder-schema==4.15.0
botframework-connector==4.15.0
botframework-streaming==4.15.0
boto3==1.34.70
botocore==1.34.70
cachetools==5.3.3
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.1
confection==0.1.4
contourpy==1.2.0
cryptography==3.4.8
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.3
dataclasses-json-speakeasy==0.5.11
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
Django==4.2.9
django-cors-headers==4.3.1
django-extensions==3.2.3
django_csp==3.8
dnspython==2.5.0
doc2text==0.2.4
docx2txt==0.8
ecdsa==0.18.0
emoji==2.11.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
et-xmlfile==1.1.0
eventlet==0.35.1
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
falcon==3.1.3
fastapi==0.109.0
fastjsonschema==2.19.1
filelock==3.13.1
filetype==1.2.0
fonttools==4.47.2
forex-python==1.8
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
google-auth==2.28.1
google-auth-oauthlib==1.2.0
greenlet==3.0.3
gspread==6.0.2
gunicorn==21.2.0
h11==0.14.0
html5==0.0.9
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.18.1
ipywidgets==8.1.1
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==1.4.2
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
langchain==0.1.0
langchain-community==0.0.11
langchain-core==0.1.10
langchain-experimental==0.0.49
langchain-openai==0.0.2.post1
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.0.85
lxml==5.2.1
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mime==0.1.0
mistune==3.0.2
mongoengine==0.27.0
mpmath==1.3.0
msal==1.28.0
msrest==0.7.1
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.3
oauthlib==3.2.2
openai==1.9.0
openpyxl==3.1.2
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
passlib==1.7.4
pillow==10.2.0
platformdirs==4.2.0
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.1
pyparsing==3.1.1
pypdf==4.0.1
PyPDF2==3.0.1
pyspellchecker==0.8.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.2.7
python-jose==3.3.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.4
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.8.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.3
starlette==0.35.1
StrEnum==0.4.15
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.0
thinc==8.2.2
threadpoolctl==3.2.0
tiktoken==0.5.2
tinycss2==1.2.1
tokenizers==0.15.1
tomli==2.0.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.2
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
unstructured==0.13.3
unstructured-client==0.18.0
uri-template==1.3.0
urllib3==1.26.18
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0